move single_node_chain to context_managers
